requirements.txt
fastapi[standard]
uvicorn
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
sqlmodel
pydantic
python-dotenv
asyncpg
psycopg2-binary
fastapi-pagination
pytest
pytest-asyncio
httpx
langchain
langchain-community
langchain-openai
openai
langchain-groq
groq
chromadb
sqlalchemy
psycopg2-binary
pypdf
transformers<5.0.0
huggingface-hub<1.0.0
sentence-transformers
langchain-huggingface
python-multipart